refactor(team_hub): #736 state.rs god-file を sub-module 分割し god-fn を段階関数化
`src-tauri/src/team_hub/state.rs` (2485 行) が HubState 14 フィールドと
TeamHub impl 全体を抱える god-file になっていたため、責務ごとの sub-module に分割。
また `team_send` (534 行 god-fn) を段階関数に分解した。振る舞いは一切変えていない
純粋なリファクタで、lock の取得/解放タイミング・メッセージ順序・エラー挙動を保持する。

state.rs → team_hub/state/ 配下に分割:
- mod.rs            — モジュール宣言 + re-export ハブ
- hub_state.rs      — HubState struct + in-memory 型群 + コンストラクタ + start/info
- recruit.rs        — recruit / pending / handshake grant / ack / semaphore 型 + impl
- member_diagnostics.rs — MemberDiagnostics struct + diagnostic 計算 impl
- file_locks_glue.rs    — file_locks / dynamic role / engine policy 連携 impl
- persistence.rs    — register_team / clear_team / persist / handoff lifecycle

team_send を段階関数に分解 (protocol/tools/send.rs 内):
parse_send_args → spool_oversized_message → resolve_send_targets →
insert_team_message → dispatch_injects → build_send_response
ロック再取得 (delivered_to/delivered_at 更新) は MessageInsertionGuard 型に
固定し、挿入済み message の再取得経路を型で表現した (新規 lock は導入しない)。

可視性変更は分割に必要な最小限のみ。#742 (handshake token TTL/single-use/
agent_id binding) の変更は recruit.rs にそのまま保持。

Closes #736
